requirements.txt
Django==3.1
djangorestframework==3.12.4   
django-rest-knox==4.1.0  
django-filter==21.1
Markdown==3.3.6
pytz==2021.3
psycopg2-binary==2.9.2
drf-yasg==1.20.0